refactor(app/test): 테스트 인프라 클래스를 coffeeshout.support.app 패키지로 이동
- coffeeshout.global.ServiceTest → coffeeshout.support.app.ServiceTest
- coffeeshout.global.StreamMockedServiceTest → coffeeshout.support.app.StreamMockedServiceTest
- coffeeshout.global.config.{IntegrationTestConfig,ServiceTestConfig} → coffeeshout.support.app.config
- coffeeshout.fixture.IntegrationTestSupport → coffeeshout.support.app.IntegrationTestSupport
- coffeeshout.fixture.WebSocketIntegrationTestSupport → coffeeshout.support.app.WebSocketIntegrationTestSupport
- 픽스처 데이터(MiniGameResultFixture, TestDataHelper)는 coffeeshout.fixture 유지
- 참조 파일 전체 import 경로 업데이트
