+++ backend/app/src/test/java/coffeeshout/global/RedisContainerIntegrationTest.java
@@ -2,7 +2,7 @@ package coffeeshout.global;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import coffeeshout.fixture.IntegrationTestSupport;
+import coffeeshout.support.app.IntegrationTestSupport;
 import java.time.Duration;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
