+++ backend/app/src/test/java/coffeeshout/report/ratelimit/ReportRateLimitStoreTest.java
@@ -3,7 +3,7 @@ package coffeeshout.report.ratelimit;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.awaitility.Awaitility.await;
 
-import coffeeshout.global.ServiceTest;
+import coffeeshout.support.app.ServiceTest;
 import io.micrometer.core.instrument.MeterRegistry;
 import java.time.Duration;
 import org.junit.jupiter.api.DisplayName;
