+++ backend/app/src/test/java/coffeeshout/report/ui/ReportControllerTest.java
@@ -5,7 +5,7 @@ import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import coffeeshout.fixture.IntegrationTestSupport;
+import coffeeshout.support.app.IntegrationTestSupport;
 import coffeeshout.report.ratelimit.ReportRateLimitStore;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.util.Map;
