+++ backend/app/src/test/java/coffeeshout/support/app/WebSocketIntegrationTestSupport.java
@@ -1,4 +1,4 @@
-package coffeeshout.fixture;
+package coffeeshout.support.app;
 
 import static org.skyscreamer.jsonassert.JSONCompareMode.LENIENT;
