+++ backend/app/src/test/java/coffeeshout/websocket/ui/WsRecoveryControllerTest.java
@@ -4,7 +4,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import coffeeshout.fixture.IntegrationTestSupport;
+import coffeeshout.support.app.IntegrationTestSupport;
 import coffeeshout.websocket.WsRecoveryService;
 import coffeeshout.websocket.StompSessionManager;
 import org.junit.jupiter.api.AfterEach;
