+++ backend/docs/adr/index.md
@@ -21,3 +21,4 @@
 | [0013](0013-domain-module-test-isolation.md)                  | 도메인 모듈 테스트 독립 실행 전략                                                           | 적용됨                   | `*/src/test/`, `:test-support`, `build.gradle.kts` (testcontainers 버전)                                                                      | TestApplication + TestConfig 패턴; `:app` 고유 빈은 모듈별 TestConfig에서 mock; 공통 설정은 `application-test-base.yml`에 집중; `CommonTestSchedulerConfig`로 공통 스케줄러 빈 중앙화; `IntegrationTestSupport`에 `@BeforeEach` DB cleanup 포함 |
 | [0014](0014-shared-web-module.md)                             | 공유 웹 인프라 모듈 `:web` 도입                                                          | 적용됨                   | `RestExceptionHandler`, `SwaggerConfig`, 전 도메인 모듈 `build.gradle.kts`, 루트 `build.gradle.kts`                                               | `:web`은 `:infra` 위에 위치; REST 엔드포인트 모듈은 `:web` 의존; `web`, `validation`, `springdoc`은 `:web`이 `api`로 노출 |
 | [0015](0015-test-base-class-inheritance.md)                   | `:test-support` 베이스 테스트 클래스 계층 설계                                               | 적용됨                   | `*/src/test/`, `:test-support`, `ServiceTest`, `IntegrationTestSupport`                                                                     | `@SpringBootTest`는 실행 모듈 클래스패스 기준; Spring `MergedAnnotations`로 상위 클래스 `@Import` 합산; 도메인 모듈은 `coffeeshout.support.ServiceTest` 직접 상속 |
+| [0016](0016-test-fixture-structure.md)                        | 테스트 픽스처 구조 — testFixtures 소스 셋과 유형별 네이밍 전략                                      | 적용됨                   | `*/src/testFixtures/`, `*/src/test/java/coffeeshout/fixture/`, `build.gradle.kts`                                                           | 공유 픽스처는 `src/testFixtures/coffeeshout.fixture`; `*Fixture`(팩토리) / `*Fake` / `*Dummy` / `Stub*` / `TestDataHelper`(@Component DB헬퍼) 유형 구분 |
